+++ backup/backup.sh
@@ -1,0 +1,3 @@
+aws s3 sync world s3://tom.25565/world
+aws s3 sync world_nether s3://tom.25565/world_nether
+aws s3 sync world_the_end s3://tom.25565/world_the_end