+++ backup/main.tf
@@ -1,0 +1,19 @@
+terraform {
+  required_providers {
+    aws = {
+      source  = "hashicorp/aws"
+      version = "~> 5.16.1"
+    }
+  }
+  required_version = "~> 1.1"
+  backend "s3" {
+    bucket         = "architectf"
+    key            = "minecraft"
+    region         = "us-east-1"
+    dynamodb_table = "architectf-timeline"
+  }
+}
+
+provider "aws" {
+  region = "us-east-1"
+}