+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 saves to the blocks aren't tracked here. just configurations.
 
+- build: backup world saves to a safe bucket using a script 2024-08-07
 - feat: initialize a server with default settings and such 2024-08-07